did a whole lotta restructuring
